Upload momentum transport model and resultion index function object
The upload of the Lapack-based sigma model of Nicoud et al. (1), described by Method B in the Appendix, is available. Additionally, the resolution index, the corresponding function objects based on Pope (2), and the effective viscosity and effective Kolmogorov length scale based on Celik et al. (3) are added to check these.

1) Nicoud, F., Toda, H. B., Cabrit, O., Bose, S., & Lee, J. (2011). Using singular values to build a subgrid-scale model for large eddy simulations. Physics of fluids, 23(8).

2) Pope, S. B. (2000). Turbulent Flows. Cambridge: Cambridge University Press.

3) Celik, I. B., Cehreli, Z. N., & Yavuz, I. (2005). Index of resolution quality for large eddy simulations.